separate art quiz buttons into diff files
